removing duplicate code
