+++ search/search.go
@@ -27,14 +27,6 @@ var reflectStaticSizeDocumentMatch int
 var reflectStaticSizeSearchContext int
 var reflectStaticSizeLocation int
 
-const SearchIOStatsCallbackKey = "_search_io_stats_callback_key"
-
-type SearchIOStatsCallbackFunc func(uint64)
-type SearchCostAggregatorCallbackFn func(string, string, uint64)
-
-const SearchCostAggregatorKey = "_search_cost_aggregator_key"
-const QueryTypeKey = "_query_type_key"
-
 func init() {
 	var dm DocumentMatch
 	reflectStaticSizeDocumentMatch = int(reflect.TypeOf(dm).Size())
